relocated delta helper, fixed min bug, added debounce
